vim: initial IdeaVim configuration
IdeaVim's Vimscript is not quite as capable as vanilla Vimscript, so
configs cannot be easily shared between it and vanilla Vim/Neovim. Some
of the more simple settings can be shared (located in common.vim) but
much cannot.

In addition, IdeaVim lets IDEA IDE actions be bound in Vimscript. This
part of the config is very much a work in progress as there is a lot of
functionality that could be utilized through these means.

The config located in which_key is for binding prefixed with Space and
intended to be used with the WhichKey IdeaVim plugin

https://github.com/TheBlob42/idea-which-key

Also, while not represented in this repo, this IdeaVim companion plugin
IdeaVim-EasyMotion is highly recommended.

https://github.com/AlexPl292/IdeaVim-EasyMotion