vim: initial vim/nvim config
Set up initial configuration for Vim/Neovim.

This is using VimPlug for plugin management. Right now there are not a
lot of plugins active, but that might change.

While Neovim offers a lot of great functionality, including LSP server
and Lua scripting support, in practice it seems easier to Vimify the IDE
than to IDEify the Vim.