vim: fix copy/paste on windows wsl
The mappings to yank into the system clipboard register (+) does not
work on Windows within the WSL. In order to get around this, added
separate configuration for Windows that will implement the same
functionality using Windows exes.

Windows provides a CLI copy utility accessible from WSL, but doesn't
have an equivalent paste utility. We can use powershell to achieve this,
but it does introduce an odd issue where the terminal font is forcibly
changed (comment in code elaborates further). This is not a showstopping
bug, however, as we can get around it by just changing the default font.